Crate datagram_socket

Macros

poll_recvmmsg

poll_sendmmsg

Structs

MaybeConnectedSocket

A cheap wrapper around a datagram socket which describes if it is connected to an explicit peer.

QuicAuditStats

SocketStats

StartupExit

Statistics from when a CCA first exited the startup phase.

Enums

StartupExitReason

The reason a CCA exited the startup phase.

StreamClosureKind

Constants

MAX_DATAGRAM_SIZE

Traits

AsDatagramSocketRecv

A convenience method that can be implemented for any type if it wants to forward its DatagramSocketRecv functionality to an inner field/socket. This automatically derives DatagramSocketRecv.

AsDatagramSocketSend

A convenience method that can be implemented for any type if it wants to forward its DatagramSocketSend functionality to an inner field/socket. This automatically derives DatagramSocketSend.

AsSocketStats

DatagramSocket

Describes an implementation of a connected datagram socket.

DatagramSocketRecv

Describes the receive half of a connected datagram socket.

DatagramSocketRecvExt

Reads datagrams from a socket.

DatagramSocketSend

Describes the send half of a connected datagram socket.

DatagramSocketSendExt

Writes datagrams to a socket.

DatagramSocketWithStats

ShutdownConnection

ShutdownConnectionExt

Shuts down a datagram oriented connection.

Functions

is_nonblocking

recvmmsg

sendmmsg