datagram_socket

Struct MaybeConnectedSocket

pub struct MaybeConnectedSocket<T> { /* private fields */ }

A cheap wrapper around a datagram socket which describes if it is connected to an explicit peer.

This struct essentially forwards its underlying socketโs send_to() method to send() if the socket is explicitly connected to a peer. This is helpful for preventing issues on platforms that do not support send_to on already-connected sockets.

Warning

A socketโs โconnectednessโ is determined once, when it is created. If the socket is created as connected, then later disconnected from its peer, its send_to() call will fail.

For example, MacOS errors if send_to is used on a socket thatโs already connected. Only send can be used. By using MaybeConnectedSocket, you can use the same send and send_to APIs in both client- and server-side code. Clients, usually with connected sockets, will then forward send_to to send, whereas servers, usually with unconnected sockets, will use send_to.

Implementations

impl<T: DatagramSocketSend> MaybeConnectedSocket<T>

pub fn new(inner: T) -> Self

pub fn inner(&self) -> &T

Provides access to the wrapped socket, allowing the user to override send_to() behavior if required.

pub fn into_inner(self) -> T

Consumes self, returning the wrapped socket.