datagram_socket

Trait AsDatagramSocketRecv

pub trait AsDatagramSocketRecv {
    type AsRecv: DatagramSocketRecv + ?Sized;

    // Required methods
    fn as_datagram_socket_recv(&mut self) -> &mut Self::AsRecv;
    fn as_shared_datagram_socket_recv(&self) -> &Self::AsRecv;
}

A convenience method that can be implemented for any type if it wants to forward its DatagramSocketRecv functionality to an inner field/socket. This automatically derives DatagramSocketRecv.

Required Associated Types

type AsRecv: DatagramSocketRecv + ?Sized

Required Methods

fn as_datagram_socket_recv(&mut self) -> &mut Self::AsRecv

fn as_shared_datagram_socket_recv(&self) -> &Self::AsRecv