datagram_socket

Trait AsDatagramSocketSend

pub trait AsDatagramSocketSend {
    type AsSend: DatagramSocketSend + ?Sized;

    // Required method
    fn as_datagram_socket_send(&self) -> &Self::AsSend;
}

A convenience method that can be implemented for any type if it wants to forward its DatagramSocketSend functionality to an inner field/socket. This automatically derives DatagramSocketSend.

Required Associated Types

type AsSend: DatagramSocketSend + ?Sized

Required Methods

fn as_datagram_socket_send(&self) -> &Self::AsSend