datagram_socket

Trait ShutdownConnectionExt

pub trait ShutdownConnectionExt: ShutdownConnection {
    // Provided method
    fn shutdown_connection(&mut self) -> impl Future<Output = Result<()>> { ... }
}

Shuts down a datagram oriented connection.

Implemented as an extension trait, adding utility methods to all ShutdownConnection types. Callers will tend to import this trait instead of ShutdownConnection.

Provided Methods

fn shutdown_connection(&mut self) -> impl Future<Output = Result<()>>

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.